PlaTo-83: refactor
